remove x and ncurses,
Removing the kb_scanner that uses X because frankly it is bad. I can bring back the code later if I want a floating X kb grabber for the process. But for now, the tty and pty kb drivers are going to work just fine.

Also removing ncurses because it cannot handle the utf8 and kb events that I need.